moved get_elevation to Map
